agents_prototype/pyproject.toml
[project]
name = "india-ai-ner"
version = "0.1.0"
description = "Agentic document ingestion, OCR, and NER pipeline using Google ADK"
authors = [
    { name = "Karthikeyan TS", email = "karthikeyan@example.com" },
]
requires-python = ">=3.10,<3.14"

dependencies = [
    # Core ADK + infra
    "google-adk>=1.15.0,<2.0.0",
    "google-cloud-aiplatform[evaluation,agent-engines]>=1.118.0,<2.0.0",
    "google-cloud-logging>=3.12.0,<4.0.0",
    "opentelemetry-instrumentation-google-genai>=0.1.0,<1.0.0",
    "gcsfs>=2024.11.0",
    "protobuf>=4.25.3",
    # Document ingestion & vision
    "pillow",
    "pypdfium2",
    "opencv-python-headless",
    "numpy",
    # LLM / config
    "google-generativeai",
    "python-dotenv",
    "easyocr>=1.7.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.23.8,<1.0.0",
    "nest-asyncio>=1.6.0,<2.0.0",
]

ocr = [
    # Heavy deps isolated on purpose
    "easyocr",
]

[project.optional-dependencies]
jupyter = [
    "jupyter>=1.0.0,<2.0.0",
]

lint = [
    "ruff>=0.4.6,<1.0.0",
    "mypy>=1.15.0,<2.0.0",
    "codespell>=2.2.0,<3.0.0",
    "types-pyyaml>=6.0.12.20240917,<7.0.0",
    "types-requests>=2.32.0.20240914,<3.0.0",
]

# -------------------------------------------------------------------
# Tooling configuration (unchanged, already good)
# -------------------------------------------------------------------

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E","F","W","I","C","B","UP","RUF"]
ignore = ["E501","C901","B006"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc","no-untyped-call","no-any-return"]
exclude = [".venv"]

[tool.codespell]
ignore-words-list = "rouge"
skip = "./locust_env/*,uv.lock,.venv,./frontend,**/*.ipynb"

[tool.pytest.ini_options]
pythonpath = "."
asyncio_default_fixture_loop_scope = "function"

# -------------------------------------------------------------------
# Build system
# -------------------------------------------------------------------

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# -------------------------------------------------------------------
# ADK / Agent Starter metadata (leave as-is, good)
# -------------------------------------------------------------------

[tool.agent-starter-pack]
name = "adk-base-3"
description = "A base ReAct agent built with Google's Agent Development Kit (ADK)"
base_template = "adk_base"
agent_directory = "app"

generated_at = "2025-12-14T16:06:11.547530+00:00"
asp_version = "0.28.2"

[tool.agent-starter-pack.create_params]
deployment_target = "agent_engine"
session_type = "none"
cicd_runner = "skip"
include_data_ingestion = true
datastore = "none"
frontend_type = "None"
